reduce ping-ponging of image load requests
