+++ maxima-ui/src/bridge/mod.rs
@@ -1,5 +1,4 @@
 pub mod game_details;
-pub mod game_images;
 pub mod get_friends;
 pub mod get_games;
 pub mod login_oauth;
